refactor: (#345) 클래스 분리
