+++ dms-infrastructure/src/main/kotlin/team/aliens/dms/persistence/student/repository/StudentTagJpaRepository.kt
@@ -1,7 +1,7 @@
 package team.aliens.dms.persistence.student.repository
 
 import org.springframework.data.repository.CrudRepository
+import team.aliens.dms.persistence.student.entity.StudentTagId
 import team.aliens.dms.persistence.student.entity.StudentTagJpaEntity
-import java.util.UUID
 
-interface StudentTagJpaRepository : CrudRepository<StudentTagJpaEntity, UUID>
+interface StudentTagJpaRepository : CrudRepository<StudentTagJpaEntity, StudentTagId>
